requirements.txt
Flask>=3.1
Flask-SQLAlchemy>=3.1
Flask-Login>=0.6
python-dotenv>=1.0
pypdf>=3.0
cryptography>=43.0
pytest>=7.0
pytest-flask>=1.2